User 컴포넌트 전체 refactoring
